Update text related ui test
