--- Example/OpenSwiftUIUITests/View/Text/TextFormatStyleUITests.swift
@@ -1,16 +1,0 @@
-//
-//  TextFormatStyleUITests.swift
-//  OpenSwiftUIUITests
-
-import SnapshotTesting
-import Testing
-@testable import TestingHost
-
-@MainActor
-@Suite(.snapshots(record: .never, diffTool: diffTool))
-struct TextFormatStyleUITests {
-    @Test(.disabled("Text layout is not ready"))
-    func dateFormatStyleExample() {
-        openSwiftUIAssertSnapshot(of: TextFormatStyleExample())
-    }
-}
